feat(users)!: remove legacy v1 user handling
Removes the deprecated v1 user endpoints and Musician-based HMAC
authentication now that all clients use ASP.NET Core Identity (v2):

- Delete UsersV1Controller, IUserRepository/UserRepository (HMAC auth
  against Musician) and the UserGroup entity/table.
- Simplify LegacyAuthResolver -> AuthResolver: it now only resolves an
  ApplicationUser by id, with no v1-token/Musician fallback.
- Trim Musician down to Id/Name/ApplicationUserId/MusicianMusicParts -
  it is now purely the part-assignment join entity, no longer an auth
  artifact. Drops the obsolete Email/PasswordHash/PasswordSalt/
  Inactive/UserGroupId columns and the UserGroups table via migration.
- Remove the ApiUser(Musician) constructor overload.

BREAKING CHANGE: the default-version (v1) /token, /users/register,
/users and /users/{id} endpoints are gone. Clients must send
x-api-version: 2.0 and authenticate via ASP.NET Core Identity.

Closes #273
