+++ src/SheetMusic.Api/Configuration/IServiceCollectionExtensions.cs
@@ -136,7 +136,7 @@ public static class IServiceCollectionExtensions
 
                     if (Guid.TryParse(context.Principal?.Identity?.Name, out var userId))
                     {
-                        var resolver = context.HttpContext.RequestServices.GetRequiredService<LegacyAuthResolver>();
+                        var resolver = context.HttpContext.RequestServices.GetRequiredService<AuthResolver>();
                         var resolved = await resolver.ResolveAsync(userId);
 
                         if (resolved is null || resolved.IsInactive)
@@ -179,7 +179,7 @@ public static class IServiceCollectionExtensions
             options.AddPolicy(AuthPolicy.Admin, policy => policy.RequireRole(Roles.Admin));
             options.AddPolicy(AuthPolicy.ManageMusic, policy => policy.RequireRole(Roles.Admin, Roles.Noteansvarlig));
         });
-        services.AddScoped<LegacyAuthResolver>();
+        services.AddScoped<AuthResolver>();
 
         return services;
     }
